all database sensor tests: text null data

+++ dqops/src/integration-test/java/com/dqops/duckdb/sensors/column/text/DuckdbColumnTextTextLengthAboveMaxLengthCountSensorParametersSpecIntegrationTest.java
@@ -74,8 +74,7 @@ public class DuckdbColumnTextTextLengthAboveMaxLengthCountSensorParametersSpecIn
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        // no option to set no params, check the SensorParametersSpec
-        Assertions.assertEquals(0.0, ValueConverter.toDouble(resultTable.column(0).get(0)));
+        Assertions.assertEquals(0L, ValueConverter.toLong(resultTable.column(0).get(0)));
     }
 
     @Test
@@ -96,7 +95,7 @@ public class DuckdbColumnTextTextLengthAboveMaxLengthCountSensorParametersSpecIn
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(0.0, ValueConverter.toDouble(resultTable.column(0).get(0)));
+        Assertions.assertEquals(0L, ValueConverter.toLong(resultTable.column(0).get(0)));
     }
 
     @Test
